requirements.txt
streamlit==1.47.1
transformers==4.49.0
torch==2.4.1     
pandas==2.3.1
openpyxl==3.1.5
scikit-learn==1.4.2
spacy==3.8.7
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
matplotlib==3.9.0
nltk==3.9.1
fpdf2==2.8.4
wordcloud==1.9.3
langdetect==1.0.9
rake-nltk==1.0.6
seaborn==0.13.2
plotly==5.23.0
gradio==4.44.1
bcrypt
pyotp
qrcode
psycopg2-binary
